fix: onboard, GET_to_POST

+++ src/onboard/onboard.controller.ts
@@ -18,7 +18,7 @@ export class OnboardController {
     @ApiBody({type: onboardDto})
     @UseGuards(JwtServiceAuthGuard)
     async setOnboard(@Body() body, @Req() req, @Res() res) {
-        let onboard = await this.onboardService.find(req);
+        let onboard = await this.onboardService.findOnboard(req);
         await this.onboardService.saveData(body);
         let info: string = (onboard = null) ? "successfully created" : "successfully changed";
         res.status(HttpStatus.CREATED).json({
@@ -29,20 +29,24 @@ export class OnboardController {
     @ApiBearerAuth()
     @ApiOperation({summary: 'onboard 조회', description: '사용자의 onboard를 조회한다. '})
     @ApiResponse({ status: 200, description: 'success', type: onboardDto})
-    @Get("checkOnboard")
+    @Post("checkOnboard")
     @UseGuards(JwtServiceAuthGuard)
     async checkOnboard(@Req() req, @Res() res){
-        let onboard = await this.onboardService.find(req);
+        let onboard = await this.onboardService.findOnboard(req);
         res.status(HttpStatus.OK).json(onboard)
     }
 
     @ApiBearerAuth()
     @ApiOperation({summary: 'onboard 삭제', description: '사용자의 onboard를 삭제한다. '})
     @ApiResponse({ status: 200, description: 'success - 삭제된 온보딩', type: onboardDto})
+    @ApiResponse({status: 404, description: 'failed - 온보딩 없음'})
     @Delete("removeOnboard")
     @UseGuards(JwtServiceAuthGuard)
     async deleteOnboard(@Req() req, @Res() res){
-        let onboard = await this.onboardService.find(req);
+        let onboard = await this.onboardService.findOnboard(req);
+        if (onboard == null) {
+            return res.status(HttpStatus.NOT_FOUND).json({});
+        }
         let del = await this.onboardService.deleteData(onboard);
         res.status(HttpStatus.OK).json(del)
     }
